Mask R-CNN/requirements.txt
Pillow==7.2
numpy==1.18.5
tensorflow>=2.2.0
opencv-python>=4.2.0
scipy==1.4.1
scikit-image==0.17.2
matplotlib==3.3.2
IPython[all]